+++ src/vfs/sahpool.rs
@@ -39,7 +39,9 @@ use crate::vfs::utils::{
     check_import_db, random_name, register_vfs, registered_vfs, ImportDbError, RegisterVfsError,
     SQLiteIoMethods, SQLiteVfs, VfsAppData, VfsError, VfsFile, VfsResult, VfsStore,
 };
+use std::cell::Cell;
 
+use crate::utils::SQLiteVfsFile;
 use js_sys::{Array, DataView, IteratorNext, Map, Reflect, Set, Uint8Array};
 use wasm_bindgen::{JsCast, JsValue};
 use wasm_bindgen_futures::JsFuture;
@@ -91,6 +93,10 @@ struct OpfsSAHPool {
     map_filename_to_sah: Map,
     /// Maps SAHs to their opaque file names
     map_sah_to_opaque_name: Map,
+    is_paused: Cell<bool>,
+    open_files: Set,
+    /// Sqlite3 VFS and makeDflt argument
+    vfs: Cell<(*mut sqlite3_vfs, bool)>,
 }
 
 impl OpfsSAHPool {
@@ -146,6 +152,9 @@ impl OpfsSAHPool {
             map_filename_to_sah: Map::new(),
             available_sah: Set::default(),
             map_sah_to_opaque_name: Map::new(),
+            is_paused: Cell::new(false),
+            open_files: Set::default(),
+            vfs: Cell::new((std::ptr::null_mut(), false)),
         };
 
         pool.acquire_access_handles(clear_files).await?;
@@ -273,6 +282,7 @@ impl OpfsSAHPool {
                 .fill(0, filename.len() as u32, HEADER_MAX_FILENAME_SIZE as u32);
             self.map_filename_to_sah.set(&JsValue::from(filename), sah);
             self.available_sah.delete(sah);
+            self.open_files.add(&JsValue::from(filename));
         } else {
             self.header_buffer
                 .fill(0, 0, HEADER_MAX_FILENAME_SIZE as u32);
@@ -363,6 +373,83 @@ impl OpfsSAHPool {
             .ok()
             .filter(|x| !x.done())
             .map(|x| x.value().into())
+    }
+
+    /// "Pauses" this VFS by unregistering it from SQLite and
+    /// relinquishing all open SAHs, leaving the associated files
+    /// intact. If this instance is already paused, this is a
+    /// no-op. Returns a Result.
+    ///
+    /// This method returns an error if SQLite has any opened file handles
+    /// hosted by this VFS, as the alternative would be to invoke
+    /// Undefined Behavior by closing file handles out from under the
+    /// library. Similarly, automatically closing any database handles
+    /// opened by this VFS would invoke Undefined Behavior in
+    /// downstream code which is holding those pointers.
+    ///
+    /// If this method returns and error due to open file handles then it has
+    /// no side effects. If the OPFS API returns an error while closing handles
+    /// then the VFS is left in an undefined state.
+    fn pause_vfs(&self) -> Result<()> {
+        if self.is_paused.get() {
+            return Ok(());
+        }
+
+        if self.open_files.size() > 0 {
+            return Err(OpfsSAHError::Generic(
+                "Cannot pause: files may be in use".to_string(),
+            ));
+        }
+
+        if self.get_capacity() > 0 {
+            let vfs = self.vfs.get();
+            if !vfs.0.is_null() {
+                unsafe {
+                    sqlite3_vfs_unregister(vfs.0);
+                }
+            }
+            self.release_access_handles();
+        }
+
+        self.is_paused.set(true);
+
+        Ok(())
+    }
+
+    /// "Unpauses" this VFS, reacquiring all SAH's and (if successful)
+    /// re-registering it with SQLite. This is a no-op if the VFS is
+    /// not currently paused.
+    ///
+    /// The returned Future resolves to a Result. See
+    /// acquire_access_handles() for how it behaves if it returns an error due to
+    /// SAH acquisition failure.
+    async fn unpause_vfs(&self) -> Result<()> {
+        if !self.is_paused.get() {
+            return Ok(());
+        }
+
+        self.acquire_access_handles(false).await?;
+
+        let vfs = self.vfs.get();
+        if vfs.0.is_null() {
+            return Err(OpfsSAHError::Generic(
+                "VFS pointer is null. Did you forget to install?".to_string(),
+            ));
+        }
+
+        match unsafe { sqlite3_vfs_register(vfs.0, i32::from(vfs.1)) } {
+            SQLITE_OK => {
+                self.is_paused.set(false);
+                Ok(())
+            }
+            error_code => Err(OpfsSAHError::Generic(format!(
+                "Failed to register VFS (SQLite error code: {error_code}"
+            ))),
+        }
+    }
+
+    fn close_file(&self, file: &mut str) {
+        self.open_files.delete(&JsValue::from(&*file));
     }
 
     fn export_db(&self, filename: &str) -> Result<Vec<u8>> {
@@ -575,6 +662,23 @@ impl SQLiteIoMethods for SyncAccessHandleIoMethods {
     ) -> ::std::os::raw::c_int {
         SQLITE_IOCAP_UNDELETABLE_WHEN_OPEN
     }
+
+    unsafe extern "C" fn xClose(pFile: *mut sqlite3_file) -> ::std::os::raw::c_int {
+        let vfs_file = SQLiteVfsFile::from_file(pFile);
+        let app_data = Self::Store::app_data(vfs_file.vfs);
+
+        app_data.close_file(vfs_file.name());
+
+        if vfs_file.flags & SQLITE_OPEN_DELETEONCLOSE != 0 {
+            if let Err(err) = Self::Store::delete_file(vfs_file.vfs, vfs_file.name()) {
+                return app_data.store_err(err);
+            }
+        }
+
+        drop(Box::from_raw(vfs_file.name()));
+
+        SQLITE_OK
+    }
 }
 
 struct SyncAccessHandleVfs;
@@ -783,6 +887,18 @@ impl OpfsSAHPoolUtil {
     pub fn count(&self) -> u32 {
         self.pool.get_file_count()
     }
+
+    pub fn pause_vfs(&self) -> Result<()> {
+        self.pool.pause_vfs()
+    }
+
+    pub async fn unpause_vfs(&self) -> Result<()> {
+        self.pool.unpause_vfs().await
+    }
+
+    pub fn is_paused(&self) -> bool {
+        self.pool.is_paused.get()
+    }
 }
 
 /// Register `opfs-sahpool` vfs and return a management tool which can be used
@@ -794,17 +910,17 @@ pub async fn install(options: &OpfsSAHPoolCfg, default_vfs: bool) -> Result<Opfs
     static REGISTER_GUARD: tokio::sync::Mutex<()> = tokio::sync::Mutex::const_new(());
     let _guard = REGISTER_GUARD.lock().await;
 
-    let vfs = if let Some(vfs) = registered_vfs(&options.vfs_name)? {
-        vfs
-    } else {
-        register_vfs::<SyncAccessHandleIoMethods, SyncAccessHandleVfs>(
+    let vfs = match registered_vfs(&options.vfs_name)? {
+        Some(vfs) => vfs,
+        None => register_vfs::<SyncAccessHandleIoMethods, SyncAccessHandleVfs>(
             &options.vfs_name,
             OpfsSAHPool::new(options).await?,
             default_vfs,
-        )?
+        )?,
     };
 
     let pool = unsafe { SyncAccessHandleStore::app_data(vfs) };
+    pool.vfs.set((vfs, default_vfs));
 
     Ok(OpfsSAHPoolUtil { pool })
 }